+++ Build.csproj
@@ -12,7 +12,6 @@
     <ProjectReference Include="projects/Test/Common/Common.csproj" />
     <ProjectReference Include="projects/Test/Applications/CreateChannel/CreateChannel.csproj" />
     <ProjectReference Include="projects/Test/Applications/MassPublish/MassPublish.csproj" />
-    <ProjectReference Include="projects/Test/AsyncIntegration/AsyncIntegration.csproj" />
     <ProjectReference Include="projects/Test/Integration/Integration.csproj" />
     <ProjectReference Include="projects/Test/SequentialIntegration/SequentialIntegration.csproj" />
     <ProjectReference Include="projects/Test/Unit/Unit.csproj" />
